remove Lambda class, replace TensorIterator body with ngraph::Function
