+++ ngraph/core/src/op/tensor_iterator.cpp
@@ -33,13 +33,6 @@ constexpr DiscreteTypeInfo op::v0::TensorIterator::InvariantInputDescription::ty
 constexpr DiscreteTypeInfo op::v0::TensorIterator::BodyOutputDescription::type_info;
 constexpr DiscreteTypeInfo op::v0::TensorIterator::ConcatOutputDescription::type_info;
 
-constexpr DiscreteTypeInfo op::v0::TensorIterator::BodyLambda::type_info;
-
-bool op::v0::TensorIterator::BodyLambda::visit_attributes(AttributeVisitor& visitor)
-{
-    return true;
-}
-
 op::v0::TensorIterator::TensorIterator(const OutputVector& values)
     : op::util::FusedOp(values)
 {
@@ -310,12 +303,7 @@ namespace ngraph
 
 bool op::v0::TensorIterator::visit_attributes(AttributeVisitor& visitor)
 {
-    if (!m_body)
-    {
-        m_body = make_shared<BodyLambda>();
-    }
-    shared_ptr<Lambda> lambda = m_body;
-    visitor.on_attribute("body", lambda);
+    visitor.on_attribute("body", m_body);
     visitor.on_attribute("input_descriptions", m_input_descriptions);
     visitor.on_attribute("output_descriptions", m_output_descriptions);
 
@@ -663,8 +651,7 @@ std::shared_ptr<Node>
     auto func = std::make_shared<Function>(m_body->get_results(), m_body->get_parameters());
     auto spec_func =
         specialize_function(func, types, new_shapes, std::vector<void*>(new_args.size(), nullptr));
-    op->m_body =
-        std::make_shared<BodyLambda>(spec_func->get_results(), spec_func->get_parameters());
+    op->m_body = std::make_shared<Function>(spec_func->get_results(), spec_func->get_parameters());
 
     for (auto& input_description : m_input_descriptions)
     {
